+++ packages/waku/src/router/server.ts
@@ -1,2 +1,2 @@
-export { defineRouter as unstable_defineRouter } from './defineRouter.js';
-export { createPages } from './createPages.js';
+export { defineRouter as unstable_defineRouter } from './define-router.js';
+export { createPages } from './create-pages.js';
